Updated project structure and included multiple unmarshaling techniques
