+++ README.md
@@ -1,0 +1,7 @@
+# Customer Marshaller
+
+The biggest waste of time to test branching and Go marshalling of JSON and YAML files. 
+
+Utilizes the file extensions of the filenames given to return the list of Customers stored in the file. 
+
+Gracefully handles errors BY PANICKING OMG IT'S BROKEN RUN AWAY
