--- index.html
@@ -1,1 +1,0 @@
-<h1>This is the title of this project webpage</h1>
